I'm not a chick magnet, I'm a scammer magnet.
Just got a random invite from a random scammer, advertising a website through his description, and he used a shorten twitter link to bypass links block on Steam (Something that I see more, and more often these days). And, well, When you see that this short link, redirect to a blogspot website, which is itself another redirector to an extremely young website using the same website gambling template as tons of others. That's bad for them.